Refactor: 좋아요 기능 controller 이동
