+++ src/main/java/com/beautymeongdang/domain/review/service/LikeService.java
@@ -1,0 +1,15 @@
+package com.beautymeongdang.domain.review.service;
+
+import com.beautymeongdang.domain.review.dto.CreateLikeRequestDto;
+import com.beautymeongdang.domain.review.dto.CreateLikeResponseDto;
+import com.beautymeongdang.domain.review.dto.DeleteLikeRequestDto;
+import com.beautymeongdang.domain.review.dto.DeleteLikeResponseDto;
+
+public interface LikeService {
+
+    // 좋아요 생성
+    CreateLikeResponseDto createLike(Long customerId, CreateLikeRequestDto request);
+
+    // 좋아요 취소
+    DeleteLikeResponseDto deleteLike(Long customerId, DeleteLikeRequestDto request);
+}
