+++ src/main/java/com/beautymeongdang/domain/review/service/ReviewService.java
@@ -8,12 +8,6 @@ import java.util.List;
 
 public interface ReviewService {
 
-    // 좋아요 생성
-    CreateLikeResponseDto createLike(Long customerId, CreateLikeRequestDto request);
-
-    // 좋아요 취소
-    DeleteLikeResponseDto deleteLike(Long customerId, DeleteLikeRequestDto request);
-
     // 리뷰 작성
     CreateReviewResponseDto createReview(CreateReviewRequestDto requestDto, List<MultipartFile> images);
